refactor: pull into_utc into atuin-server-database crate
